backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
async-timeout==5.0.1
attrs==25.4.0
bcrypt==5.0.0
bidict==0.23.1
boto3==1.42.0
botocore==1.41.6
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click
colorama==0.4.6
cryptography==46.0.3
Deprecated==1.3.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
gTTS
email-validator==2.3.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
fastapi==0.123.0
frozenlist==1.8.0
greenlet==3.2.4
groq==0.36.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
img2pdf==0.6.3
jiter==0.12.0
jmespath==1.0.1
langdetect==1.0.9
lxml==6.0.2
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.7.0
ocrmypdf==16.12.0
openai==2.8.1
openpyxl==3.1.5
packaging==25.0
passlib==1.7.4
pdf2image==1.17.0
pdfminer.six==20251107
pi_heif==1.1.1
pikepdf==10.0.2
pillow==12.0.0
pluggy==1.6.0
propcache==0.4.1
psycopg==3.2.13
psycopg-binary==3.2.13
psycopg2==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyPDF2==3.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.12.3
python-jose==3.5.0
python-multipart==0.0.20
python-socketio==5.15.0
PyYAML==6.0.3
reportlab==4.4.5
rich==14.2.0
rsa==4.9.1
s3transfer==0.16.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.50.0
tabulate==0.9.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn
watchfiles==1.1.1
websockets==15.0.1
wrapt==2.0.1
wsproto==1.3.2
yarl==1.22.0
google-cloud-texttospeech
google-api-core
gunicorn
uvicorn[standard]
gunicorn==22.0.0
google-cloud-vision
prometheus-client